// src/api/models/getVisitCohortItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface GetVisitCohortItem {
  /** 주차 */
  week: string;
  /** 1회 이상 방문 */
  firstVisit: number;
  /** 2회 이상 방문 */
  secondVisit: number;
  /** 3회 이상 방문 */
  thirdVisit: number;
  /** 4회 이상 방문 */
  forthVisit: number;
}